#687 | refactor(masonry): split stores to separate files
